Use dedicated module for parameters, rename functions

+++ molalignlib/biasing.f90
@@ -10,12 +10,12 @@ contains
 subroutine setadjbias(natom, nblock, blocksize, coords0, coords1, bias)
     integer, intent(in) :: natom, nblock
     integer, dimension(:), intent(in) :: blocksize
-    real, dimension(:, :), intent(in) :: coords0, coords1
-    real, dimension(:, :), intent(out) :: bias
+    real(wp), dimension(:, :), intent(in) :: coords0, coords1
+    real(wp), dimension(:, :), intent(out) :: bias
 
     integer h, i, j, offset
-!    real d0(natom, natom), d1(natom, natom)
-    real, allocatable :: d0(:, :), d1(:, :)
+!    real(wp) d0(natom, natom), d1(natom, natom)
+    real(wp), allocatable :: d0(:, :), d1(:, :)
     allocate(d0(natom, natom), d1(natom, natom))
 
     do i = 1, natom
